+++ src/app/appEvents.ts
@@ -19,24 +19,17 @@ export function initAppEvents() {
     });
 
     _k.app.onResize(() => {
-        const fixedSize = _k.globalOpt.width && _k.globalOpt.height;
-        if (fixedSize && !_k.globalOpt.letterbox) {
-            return;
-        }
-
         _k.canvas.width = _k.canvas.offsetWidth * _k.gfx.pixelDensity;
         _k.canvas.height = _k.canvas.offsetHeight * _k.gfx.pixelDensity;
-        _k.app.updateCanvasScale();
 
-        updateViewport();
-
-        if (!fixedSize) {
+        if (!_k.globalOpt.width || !_k.globalOpt.height) {
             _k.gfx.frameBuffer.free();
             _k.gfx.frameBuffer = new FrameBuffer(
                 _k.gfx.ggl,
                 _k.gfx.ggl.gl.drawingBufferWidth,
                 _k.gfx.ggl.gl.drawingBufferHeight,
             );
+
             _k.gfx.width = _k.gfx.ggl.gl.drawingBufferWidth
                 / _k.gfx.pixelDensity
                 / _k.globalOpt.scale;
@@ -44,6 +37,8 @@ export function initAppEvents() {
                 / _k.gfx.pixelDensity
                 / _k.globalOpt.scale;
         }
+
+        updateViewport();
     });
 
     if (_k.globalOpt.debug !== false) {