+++ src/core/engineLoop.ts
@@ -8,7 +8,6 @@ import {
 import type { Debug } from "../debug/debug";
 import { SystemPhase } from "../ecs/systems/systems";
 import type { Game } from "../game/game";
-import { drawDebug } from "../gfx/draw/drawDebug";
 import { drawFrame, transformFrame } from "../gfx/draw/drawFrame";
 import { drawLoadScreen } from "../gfx/draw/drawLoadingScreen";
 import { updateViewport } from "../gfx/viewport";
@@ -111,7 +110,6 @@ export function startEngineLoop(
                 }
 
                 drawFrame();
-                if (gopt.debug !== false) drawDebug();
 
                 for (const sys of game.systemsByEvent[SystemPhase.AfterDraw]) {
                     sys.run();