+++ CHANGELOG.md
@@ -66,6 +66,10 @@ So your change should look like:
   time (#1082) - @nojaf
 - Fixed mouse coordinates not being calculated properly when canvas is resized
   by CSS and wasn't rendered at its natural size (#1096) - @Stanko
+- Fixed touch coordinates in aspect-fit fullscreen with fixed canvas size and
+  letterbox disabled (#1106) - @imaginarny
+- Fixed incorrect resolution and scale of the debug inspect mode (#1106) -
+  @imaginarny
 - Modified `pos`, `skew` and `scale` components to make operations like
   `obj.pos.x += 1` work again (#1109) - @ErikGXDev
 - Fixed `scale`, `skew`, and `rotate` component transforms not being applied on